Update Method and Evaluation APIs to be suspend functions
Unfortunately, this causes us to have to update a whole bunch of method
implementation sites as well as update tests to run as a coroutine.

Also, sort needed a bit of a workover to use an algorithm that
supports being suspended.

But despite how it looks, logically this was a pretty minor change.