The great Value refactor
Value was just a glorified Kotlin value. It originally was supposed to
be an opaque value that you had to manually convert, but in practice,
it was just too useful to be able to pull the wrapped value out of it
anyway sometimes.

So instead of all the extra typing and memory allocations, we just
decided to remove it for now and lean on a more simple API.

I may regret it later, we'll see!